study_platform/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
gunicorn==20.1.0
openpyxl==3.1.5
